fix(JSAP): fixed bug in transfer send button

+++ backend/src/routes/depositRoute.js
@@ -8,7 +8,7 @@ const databaseActions = new DatabaseActions(connection);
 
 depositRoute.post('/', async (req, res) => {
   if (!requestValidator.validateContentType(req.headers['content-type'])) {
-    res.status(415).send('Content-type must be application/json.');
+    res.status(415).json({ message: 'Content-type must be application/json.' });
     return;
   }
 
@@ -16,7 +16,7 @@ depositRoute.post('/', async (req, res) => {
   const requiredKeys = ['depositName'];
   const missingKeys = requestValidator.findMissingKeys(requiredKeys, Object.keys(req.body));
   if (missingKeys.length > 0) {
-    res.status(400).send(`Missing ${missingKeys}.`);
+    res.status(400).json({ message: `Missing ${missingKeys}.` });
     return;
   }
 
@@ -29,10 +29,10 @@ depositRoute.post('/', async (req, res) => {
       const insertedAccount = await databaseActions.getAccountByDepositId(insertedId);
       res.status(200).json(insertedAccount);
     } else {
-      res.status(422).send({ message: 'Duplicate account name, please use another name' });
+      res.status(422).json({ message: 'Duplicate account name, please use another name' });
     }
   } catch (error) {
-    res.status(500).send({ message: 'Something went wrong, please try again later.' });
+    res.status(500).json({ message: 'Something went wrong, please try again later.' });
   }
 });
 
@@ -41,7 +41,7 @@ depositRoute.get('/', async (req, res) => {
     const accounts = await databaseActions.getAccountsByUserId(req.userId);
     res.status(200).json(accounts);
   } catch (error) {
-    res.status(500).send({ message: 'Something went wrong, please try again later.' });
+    res.status(500).json({ message: 'Something went wrong, please try again later.' });
   }
 });
 
@@ -51,17 +51,25 @@ depositRoute.get('/:depositId', async (req, res) => {
     const accountsDetails = await databaseActions.getTransfers(depositId, req.userId);
     res.status(200).json(accountsDetails);
   } catch (error) {
-    res.status(500).send({ message: 'Something went wrong, please try again later.' });
+    res.status(500).json({ message: 'Something went wrong, please try again later.' });
   }
 });
 
 depositRoute.delete('/:depositId', async (req, res) => {
+  if (!requestValidator.validateContentType(req.headers['content-type'])) {
+    res.status(415).json({ message: 'Content-type must be application/json.' });
+    return;
+  }
   try {
     const { depositId } = req.params;
-    await databaseActions.deleteAccount(depositId);
-    res.status(200).json({ depositId });
+    if (parseInt(req.userId) === parseInt(req.body.userId)) {
+      await databaseActions.deleteAccount(depositId);
+      res.status(200).json({ depositId });
+    } else {
+      res.status(401).json({ message: 'Unauthorized access!' });
+    }
   } catch (error) {
-    res.status(500).send({ message: 'Something went wrong, please try again later.' });
+    res.status(500).json({ message: 'Something went wrong, please try again later.' });
   }
 });
